pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openclaw-guardian"
version = "1.0.0"
description = "OpenClaw 安全卫士 - 自动监控、检测和修复 OpenClaw 运行时的安全问题"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "1 号小虫子 · 严谨专业版", email = "guardian@openclaw.dev"}
]
keywords = [
    "openclaw",
    "security",
    "monitor",
    "guardian",
    "automation",
    "safety"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: No Input/Output (Daemon)",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "psutil>=5.9.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=4.0",
]

[project.scripts]
openclaw-guardian = "main:main"
guardian = "main:main"

[project.urls]
Homepage = "https://github.com/openclaw/openclaw-guardian"
Documentation = "https://github.com/openclaw/openclaw-guardian#readme"
Repository = "https://github.com/openclaw/openclaw-guardian.git"
Issues = "https://github.com/openclaw/openclaw-guardian/issues"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
